pages/requirements.txt
phidata 
streamlit
groq
yfinance